+++ problem-solver/cxx/inferenceModule/searcher/templateSearcher/TemplateSearcherInStructures.hpp
@@ -42,8 +42,6 @@ private:
 
   std::map<std::string, std::string> getTemplateLinksContent(ScAddr const & templateAddr) override;
 
-  virtual void prepareBeforeSearch();
-
   virtual bool isValidElement(ScAddr const & element) const;
 };
 }  // namespace inference